separate docker instructions from main module + improve code style
